# core/gnosis/openapi.yaml
openapi: 3.0.3
info:
  title: Gnosis API
  version: "1.0.0"
  description: >
    This API powers the Gnosis application.
    It provides endpoints for:
      - Company management (creation and API key management)
      - Updating fact extraction prompts for companies
      - Memory ingestion and retrieval
      - Health checks
servers:
  - url: http://localhost:8787/api
    description: Local development server
paths:
...
  /v1/memories:
    post:
      summary: Add a new memory
      description: >
        Adds a new memory with the supplied messages, user ID, and namespace.
        This endpoint requires company authentication.
      tags:
        - Memories
      security:
        - companyBearerAuth: []
      requestBody:
        description: Memory payload.
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddMemoryRequest"
      responses:
        "200":
          description: Memory processed successfully.
          content:
            application/json:
              schema:
                type: object
        "500":
          description: Error processing memory.
  /v1/memories/{id}:
    get:
      summary: Retrieve a memory by ID
      description: >
        Retrieves a memory's details using its unique identifier.
        This endpoint requires company authentication.
      tags:
        - Memories
      security:
        - companyBearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          description: Memory ID.
          schema:
            type: string
      responses:
        "200":
          description: Memory details.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Memory"
        "404":
          description: Memory not found.
        "500":
          description: Server error.

  # Health check endpoint
...
components:
  securitySchemes:
    adminBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API_KEY
      description: >
        Provide the admin API key as a bearer token for admin endpoints.
    companyBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API_KEY
      description: >
        Provide the company API key as a bearer token for company endpoints.
  schemas:
...
    AddMemoryRequest:
      type: object
      properties:
        messages:
          type: array
          items:
            $ref: "#/components/schemas/Message"
        userId:
          type: string
        namespace:
          type: string
      required:
        - messages
        - userId
        - namespace
    Message:
      type: object
      properties:
        role:
          type: string
          enum:
            - system
            - user
            - assistant
        content:
          type: string
      required:
        - role
        - content
    Memory:
      type: object
      properties:
        id:
          type: string
        text:
          type: string
        metadata:
          type: object
        namespace:
          type: string
      required:
        - id
        - text
        - metadata
        - namespace